refactor(kb-load): extract lib tests to sibling file
Move the single inline `#[cfg(test)] mod tests` from
primer-kb-load/src/lib.rs into a sibling src/tests.rs. lib.rs is a
crate root, so the module resolves to src/tests.rs with no mkdir and
the test module stays a descendant — `use super::*` reaches all parent
privates verbatim, zero visibility churn.

695 -> 390 lines (-305). All 12 kb-load lib tests preserved; workspace
0 failures, clippy clean, fmt clean. Pure test-only relocation, no
behaviour change.
